refactoring BotService, replace saveUser from BotService to UserService
